+++ internal/version/utils.go
@@ -2,8 +2,8 @@ package version
 
 import versionpb "github.com/c4t-but-s4d/fastad/pkg/proto/data/version"
 
-func NewVersionProto(version int32) *versionpb.Version {
+func NewVersionProto(version int) *versionpb.Version {
 	return &versionpb.Version{
-		Version: version,
+		Version: int64(version),
 	}
 }
